digielves-dev/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
csp==0.10.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.9
django-cors-headers==4.3.1
django-crontab==0.7.1
django-extensions==3.2.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.47.0
fpdf2==2.7.7
fsspec==2023.12.2
huggingface-hub==0.20.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
memory-profiler==0.61.0
msgpack==1.0.7
numpy==1.24.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypdf2==3.0.1
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
rsa==4.9
safetensors==0.4.1
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
Telethon==1.33.1
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.36.2
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
XlsxWriter==3.2.0
zope.interface==6.1
librosa
pydub
torch
openai-whisper
